ccw/requirements.txt
cffi==1.14.5
click==7.1.2
cryptography==3.4.6
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
mysql-connector-python==8.0.23
openssl-python==0.1.1
protobuf==3.15.6
pycparser==2.20
pyOpenSSL==20.0.1
six==1.15.0
Werkzeug==1.0.1
